Add options for unknown ports and MSRs
For each VirtualMachine, we could return all ones, or a random
value, or inject a #GP if an unknown MSR is accessed. For
unknown ports, we can also return all ones or a random value.

As default, we inject #GP for unknown MSRs and return all
ones for unknown ports.

Signed-off-by: Changyuan Lyu <changyuan.lv@Gmail.com>