Add address converter to VM to handle low memory
While our purpose is to make the VM's physical address space the
same as host virtual address space, because of the limitation from
macOS, virtual address below 4GiB is not accessible. So when the
VM is created, we allocate the low memory in advance and setup a
closure as a handler for further address translation.

Signed-off-by: Changyuan Lyu <changyuan.lv@Gmail.com>